Move models to submodules.
