+++ pydis_site/apps/api/models/bot/__init__.py
@@ -1,0 +1,16 @@
+from .bot_setting import BotSetting  # noqa
+from .deleted_message import DeletedMessage  # noqa
+from .documentation_link import DocumentationLink  # noqa
+from .infraction import Infraction  # noqa
+from .message import Message  # noqa
+from .message_deletion_context import MessageDeletionContext  # noqa
+from .nomination import Nomination  # noqa
+from .off_topic_channel_name import OffTopicChannelName  # noqa
+from .reminder import Reminder  # noqa
+from .role import Role  # noqa
+from .snake_fact import SnakeFact  # noqa
+from .snake_idiom import SnakeIdiom  # noqa
+from .snake_name import SnakeName  # noqa
+from .special_snake import SpecialSnake  # noqa
+from .tag import Tag  # noqa
+from .user import User  # noqa
