+++ pydis_site/apps/api/models/bot/bot_setting.py
@@ -1,0 +1,27 @@
+from django.contrib.postgres import fields as pgfields
+from django.core.exceptions import ValidationError
+from django.db import models
+
+from pydis_site.apps.api.models.utils import ModelReprMixin
+
+
+def validate_bot_setting_name(name):
+    known_settings = (
+        'defcon',
+    )
+
+    if name not in known_settings:
+        raise ValidationError(f"`{name}` is not a known setting name.")
+
+
+class BotSetting(ModelReprMixin, models.Model):
+    """A configuration entry for the bot."""
+
+    name = models.CharField(
+        primary_key=True,
+        max_length=50,
+        validators=(validate_bot_setting_name,)
+    )
+    data = pgfields.JSONField(
+        help_text="The actual settings of this setting."
+    )
